requirements.txt
azure-core==1.18.0
azure-storage-blob==12.9.0b1
azure-storage-file-datalake==12.4.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
idna==3.2
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
pycparser==2.20
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.6
